+++ koishi/plugins/rendering_helpers/__init__.py
@@ -1,9 +1,11 @@
+from .constants import *
 from .field_renderers import *
 from .solutions import *
 from .value_renderers import *
 
 
 __all__ = (
+    *constants.__all__,
     *field_renderers.__all__,
     *solutions.__all__,
     *value_renderers.__all__,