+++ koishi/plugins/rendering_helpers/constants.py
@@ -1,0 +1,3 @@
+__all__ = ()
+
+ROLE_MENTIONS_MAX = 20